Modified store.js, renamed countries folder
