+++ src/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 import continentsSlice from "./redux/continents/continentsSlice";
-import countriesSlice from "./redux/continents/countires/countriesSlice";
+import countriesSlice from "./redux/continents/countries/countriesSlice";
 
 const store = configureStore({
   reducer: {
